renamed all SF utils files to more descriptive names
