+++ project-base/storefront/components/Basic/SymfonyDebugToolbar/SymfonyDebugToolbar.tsx
@@ -1,4 +1,4 @@
-import { useRequests } from './utils';
+import { useRequests } from './symfonyDebugToolbarUtils';
 import symfonyImage from '/public/images/symfony.svg';
 import { Image } from 'components/Basic/Image/Image';
 import { Button } from 'components/Forms/Button/Button';
