+++ project-base/storefront/components/Basic/SymfonyDebugToolbar/symfonyDebugToolbarUtils.tsx
