+++ project-base/storefront/components/Blocks/SimpleNavigation/SimpleNavigationListItem.tsx
@@ -1,4 +1,4 @@
-import { getLinkType } from './utils';
+import { getLinkType } from './simpleNavigationUtils';
 import { ExtendedNextLink } from 'components/Basic/ExtendedNextLink/ExtendedNextLink';
 import { Image } from 'components/Basic/Image/Image';
 import { TIDs } from 'cypress/tids';
