+++ project-base/storefront/components/Blocks/SimpleNavigation/simpleNavigationUtils.ts
