+++ project-base/storefront/components/Layout/Footer/DeferredFooter.tsx
@@ -1,4 +1,4 @@
-import { useFooterArticles } from './utils';
+import { useFooterArticles } from './footerUtils';
 import dynamic from 'next/dynamic';
 import { useDeferredRender } from 'utils/useDeferredRender';
