+++ project-base/storefront/components/Layout/Footer/footerUtils.ts
