+++ project-base/storefront/components/Pages/CategoryDetail/categoryDetailUtils.ts
