+++ project-base/storefront/components/Pages/Search/searchUtils.ts
